smp: don't use 16-bit words for atomic accesses

+++ include/linux/smp.h
@@ -20,7 +20,7 @@ struct call_single_data {
 	struct list_head list;
 	smp_call_func_t func;
 	void *info;
-	u16 flags;
+	unsigned int flags;
 };
 
 /* total number of cpus in this system (may exceed NR_CPUS) */
